refactor: 클래스명 및 패키지 변경
